various fixes for compilation with MINGW, Bloodshed C++ and the TDB port under WIN32
Note: TDB will NOT work if Apache is running "multithreaded" instead of "multiprocess"
a multithread version of TDB is pretty easy, but needs to be written

Hset's mess up on Win32/TDB because the background scheduler creates a thread
which connects to the hset database (on linux it's OK, because client-hits are in their own process)

POssibly just give the scheduler its own database for a Win32/TDB config?

Not sure.